AJAX/ Bootstrap Exercice Done
Pretty complex exercice... Bootstrap made me.. sad. I had the feeling to work with a pretty complicated/usless/dirty html...
But i can understand the utility.

Ajax is powerfull, and i learnt a lot. I can now interact with the url params, and it was the first time for me.

It was an hard exercice for, and i have to work harder.

Code cleaner, add some comments, code shorter....

I, serusly, have a lot of troubles.. And i want to be better.